requirements.txt
numpy
pandas
pandas_ta
pandas_market_calendars
schedule
yfinance
moomoo-api